+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix app crashing after coming back to app from the background after the overdue list is loaded
 - Handle see all/see less button clicks for pending overdue list section
 - Fix views flashing in the overdue list
+- When overdue appointments are loaded or are loading, render appropriate views
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 - [In Progress: 03 Jun 2022] Implement `OverdueAppointmentListItemNew` adapter